1st Spawning Rule Added - No Trees Next To water
Added a spawning rule function "treeWaterRule" that iterates through the map and checks in the 4 coordinates next to it to make sure there isn't any water. If water is found then it is replaced with land.

Might update function in future to check all 8 blocks around tree.
Might update to intake a char that the trees should be replaced with.